[48] apply code review suggestions
