--- .sqlx/query-ba18f9564ce40802573e18b6c3706247399abb8c545c1c03718fd97261870418.json
@@ -1,17 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "UPDATE users SET handle = $1, picture_link = $2, password_hash = $3 WHERE id = $4",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text",
-        "Text",
-        "Text",
-        "Uuid"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "ba18f9564ce40802573e18b6c3706247399abb8c545c1c03718fd97261870418"
-}
