+++ src/routes/user_routes.rs
@@ -1,5 +1,4 @@
-use crate::{omni_error::OmniError, setup::AppState, tournament::Tournament, users::{photourl::PhotoUrl, roles::Role, User}};
-use argon2::{password_hash::SaltString, Argon2, PasswordHasher};
+use crate::{omni_error::OmniError, setup::AppState, users::{photourl::PhotoUrl, UserPatch, User}};
 use axum::{
     extract::{Path, State},
     http::{HeaderMap, StatusCode},
@@ -7,22 +6,13 @@ use axum::{
     routing::get,
     Json, Router,
 };
-use rand::rngs::OsRng;
 use serde::Deserialize;
-use sqlx::{query, Pool, Postgres};
 use tower_cookies::Cookies;
 use tracing::error;
 use utoipa::ToSchema;
 use uuid::Uuid;
 use serde_json::Error as JsonError;
 
-
-#[derive(Deserialize, ToSchema)]
-pub struct UserPatch {
-    pub handle: Option<String>,
-    pub picture_link: Option<PhotoUrl>,
-    pub password: Option<String>,
-}
 
 
 #[derive(Clone, Deserialize, ToSchema)]
@@ -33,216 +23,6 @@ pub struct UserWithPassword {
     pub handle: String,
     pub picture_link: Option<PhotoUrl>,
     pub password: String,
-}
-
-impl User {
-    pub async fn get_by_id(id: Uuid, pool: &Pool<Postgres>) -> Result<User, OmniError> {
-        let user =
-            sqlx::query!("SELECT handle, picture_link FROM users WHERE id = $1", id)
-                .fetch_one(pool)
-                .await?;
-
-        Ok(User {
-            id,
-            handle: user.handle,
-            picture_link: match user.picture_link {
-                Some(url) => Some(PhotoUrl::new(&url)?),
-                None => None,
-            },
-        })
-    }
-
-    pub async fn get_by_handle(
-        handle: &str,
-        pool: &Pool<Postgres>,
-    ) -> Result<User, OmniError> {
-        let user = sqlx::query!(
-            "SELECT id, picture_link FROM users WHERE handle = $1",
-            handle
-        )
-        .fetch_one(pool)
-        .await?;
-
-        Ok(User {
-            id: user.id,
-            handle: handle.to_string(),
-            picture_link: match user.picture_link {
-                Some(url) => Some(PhotoUrl::new(&url)?),
-                None => None,
-            },
-        })
-    }
-
-    pub async fn get_all(pool: &Pool<Postgres>) -> Result<Vec<User>, OmniError> {
-        let users = sqlx::query!("SELECT id, handle, picture_link FROM users")
-            .fetch_all(pool)
-            .await?
-            .iter()
-            .map(|u| {
-                Ok(User {
-                    id: u.id,
-                    handle: u.handle.clone(),
-                    picture_link: match u.picture_link.clone() {
-                        Some(url) => Some(PhotoUrl::new(&url)?),
-                        None => None,
-                    },
-                })
-            })
-            .collect::<Result<Vec<User>, OmniError>>()?;
-        Ok(users)
-    }
-    
-    pub async fn post(
-        user: User,
-        password: String,
-        pool: &Pool<Postgres>,
-    ) -> Result<User, OmniError> {
-        let pic = match &user.picture_link {
-            Some(url) => Some(url.as_str()),
-            None => None,
-        };
-        let hash = User::generate_password_hash(&password).unwrap();
-        match sqlx::query!(
-            "INSERT INTO users VALUES ($1, $2, $3, $4)",
-            &user.id,
-            &user.handle,
-            pic,
-            hash
-        )
-        .execute(pool)
-        .await
-        {
-            Ok(_) => Ok(user),
-            Err(e) => Err(e)?,
-        }
-    }
-
-
-    pub async fn patch(
-        self,
-        patch: UserPatch,
-        pool: &Pool<Postgres>,
-    ) -> Result<User, OmniError> {
-        let picture_link = match &patch.picture_link {
-            Some(url) => Some(url.clone()),
-            None => self.picture_link.clone(),
-        };
-        let updated_user = User {
-            id: self.id,
-            handle: patch.handle.clone().unwrap_or(self.handle.clone()),
-            picture_link
-        };
-        if patch.password != None {
-            self.update_user_and_change_password(&patch, pool).await?;
-        }
-        self.update_user_without_changing_password(&patch, pool).await?;
-        Ok(updated_user)
-    }
-
-    async fn update_user_and_change_password(&self, patch: &UserPatch, pool: &Pool<Postgres>) -> Result<(), OmniError> {
-        let picture_link = match &patch.picture_link {
-            Some(url) => Some(url.as_url().to_string()),
-            None => Some(self.picture_link.as_ref().unwrap().as_str().to_owned()),
-        };
-        let password_hash = User::generate_password_hash(&patch.password.as_ref().unwrap()).unwrap().clone();
-        match query!("UPDATE users SET handle = $1, picture_link = $2, password_hash = $3 WHERE id = $4",
-            patch.handle,
-            picture_link,
-            password_hash,
-            self.id
-        ).execute(pool).await {
-            Ok(_) => Ok(()),
-            Err(e) => Err(e)?,
-        }
-    }
-    
-    fn generate_password_hash(password: &str) -> Result<String, OmniError> {
-        let hash = {
-            let argon = Argon2::default();
-            let salt = SaltString::generate(&mut OsRng);
-            match argon.hash_password(password.as_bytes(), &salt) {
-                Ok(hash) => hash.to_string(),
-                Err(e) => return Err(e)?,
-            }
-        };
-        Ok(hash)
-    }
-
-    async fn update_user_without_changing_password(&self, patch: &UserPatch, pool: &Pool<Postgres>) -> Result<(), OmniError> {
-        let picture_link = match &patch.picture_link {
-            Some(url) => Some(url.as_url().to_string()),
-            None => Some(self.picture_link.as_ref().unwrap().as_str().to_owned()),
-        };
-        match query!("UPDATE users SET handle = $1, picture_link = $2 WHERE id = $3",
-            patch.handle,
-            picture_link,
-            self.id
-        ).execute(pool).await {
-            Ok(_) => Ok(()),
-            Err(e) => Err(e)?,
-        }
-    }
-    
-
-    pub async fn delete(self, connection_pool: &Pool<Postgres>) -> Result<(), OmniError> {
-        match query!("DELETE FROM users WHERE id = $1", self.id)
-            .execute(connection_pool)
-            .await
-        {
-            Ok(_) => Ok(()),
-            Err(e) => {
-                Err(e)?
-            }
-        }
-    }
-
-    // ---------- DATABASE HELPERS ----------
-    pub async fn get_roles(
-        &self,
-        tournament: Uuid,
-        pool: &Pool<Postgres>,
-    ) -> Result<Vec<Role>, OmniError> {
-        let roles_result = sqlx::query!(
-            "SELECT roles FROM roles WHERE user_id = $1 AND tournament_id = $2",
-            self.id,
-            tournament
-        )
-        .fetch_optional(pool)
-        .await?;
-
-        if roles_result.is_none() {
-            return Ok(vec![]);
-        }
-
-        let roles = roles_result.unwrap().roles;
-        let vec = match roles {
-            Some(vec) => vec
-                .iter()
-                .map(|role| serde_json::from_str(role.as_str()))
-                .collect::<Result<Vec<Role>, JsonError>>()?,
-            None => vec![],
-        };
-        Ok(vec)
-    }
-
-    pub async fn is_organizer_of_any_tournament(&self, pool: &Pool<Postgres>) -> Result<bool, OmniError> {
-        let tournaments = Tournament::get_all(pool).await?;
-        for tournament in tournaments {
-            let roles = self.get_roles(tournament.id, pool).await?;
-            if roles.contains(&Role::Organizer) {
-                return Ok(true);
-            }
-        }
-        return Ok(false);
-        
-    }
-
-    pub async fn invalidate_all_sessions(&self, pool: &Pool<Postgres>) -> Result<(), OmniError> {
-        match query!("DELETE FROM sessions WHERE user_id = $1", self.id).execute(pool).await {
-            Ok(_) => Ok(()),
-            Err(e) => Err(e)?,
-        }
-    }
 }
 
 impl From<UserWithPassword> for User {
@@ -334,12 +114,12 @@ async fn create_user(
 ) -> Result<Response, OmniError> {
     let pool = &state.connection_pool;
     let user = User::authenticate(&headers, cookies, &pool).await?;
-    if !user.is_infrastructure_admin() && !user.is_organizer_of_any_tournament(pool).await? {
+    if !user.is_infrastructure_admin() && !user.can_create_users_within_any_tournament(pool).await? {
         return Err(OmniError::UnauthorizedError);
     }
 
     let user_without_password = User::from(json.clone());
-    match User::post(user_without_password, json.password, pool).await {
+    match User::create_user(user_without_password, json.password, pool).await {
         Ok(user) => Ok(Json(user).into_response()),
         Err(e) => {
             error!("Error creating a new user: {e}");
@@ -427,7 +207,7 @@ async fn patch_user_by_id(
     }
 
     match user_to_be_patched.patch(new_user, pool).await {
-        Ok(patched_user) => Ok(Json(patched_user).into_response()),
+        Ok(patched_user) => Ok(Json(patched_user as User).into_response()),
         Err(e) => {
             error!("Error patching a user with id {}: {e}", id);
             Err(e)?
